% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{combine_footers}
\alias{combine_footers}
\title{Combines footer data from a dpa object list into a single data
frame}
\usage{
combine_footers(dpa.list)
}
\arguments{
\item{dpa.list}{A list of dpa objects, either from loading several
files using load_dpa or combined manually. Note: the list should
include only dpa objects!}
}
\value{
A tibble, combining all footer data from dpa.list
}
\description{
Given a dpa object list, this function will extract all footers
(the additional measurement data) from all dpa objects in a given
list and combine them in a single data frame/tibble.
}
\examples{
## load all files in directory
dpa.list <- load_dpa(dpa.directory = system.file("extdata", package = "densiter"))
combine_footers(dpa.list)
}
\seealso{
load_dpa, combine_data.
}
