% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dpa_detect_end}
\alias{dpa_detect_end}
\title{Detect measurement ending point automatically using changepoint
segmentation}
\usage{
dpa_detect_end(dpa, cutoff.sd = 1, return.plot = FALSE)
}
\arguments{
\item{dpa}{A dpa object, see load_dpa.}

\item{cutoff.sd}{How many standard deviations for the cutoff limit?}

\item{return.plot}{If true, will return a plot displaying segment
detection for the current dpa file.}
}
\value{
Either a row number where the actual measurement ends or
  a plot, displaying changepoint segmentation and set limits.
}
\description{
The opposite of the dpa_detect_start, it will check the mean values
of the last four segments and compare them to the cutoff limit.
Will give a warning if end not detected, which is expected on
measurements where the needle did not exit the tree on the opposite
side of the tree. See \code{return.plot = TRUE} to display the
actual process. The function is called on a dpa object and returns
either a row number of the measurement ending or a plot displaying
the segmentation and detection. The sensitivity can be adjusted
using the cutoff.sd parameter, which is an indicator on how many
standard deviations the segment mean value can be before cutting it
off.
}
\examples{
## load a single file
dpa <- load_dpa("data/test.dpa")
## get ending point
start <- dpa_detect_end(dpa, return.plot = TRUE)
## plot the end detection
dpa_detect_end(dpa, return.plot = TRUE)
}
\seealso{
dpa_detect_start, dtrim, dtriml, dtrim_s, dtriml_s
}
