% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dpa_detect_start}
\alias{dpa_detect_start}
\title{Detect measurement starting point automatically using changepoint
segmentation}
\usage{
dpa_detect_start(dpa, cutoff.sd = 1, return.plot = FALSE)
}
\arguments{
\item{dpa}{A dpa object, see load_dpa.}

\item{cutoff.sd}{How many standard deviations for the cutoff limit?}

\item{return.plot}{If true, will return a plot displaying segment
detection for the current dpa file.}
}
\value{
Either a row number where the actual measurement starts or
  a plot, displaying changepoint segmentation and set limits.
}
\description{
A typical resistance drilling measurement starts with an increase
in resistance values in between the measurement start and the
immersion of the needle in the wood. These values are not useful
when estimating density and should be removed before further
analysis. This function will detect the starting point
automatically using binary segmentation from the package
\code{changepoint}, which separates the measurement in segments
based on their mean and variance. Start is detected, when the
segment mean is outside of the cutoff limit, see \code{return.plot
= TRUE} to display the actual process. This function will only
check the mean values of the first four (4) segments and compare
them to the cutoff value. The function is called on a dpa object
and returns either a row number of the starting point or a plot
displaying the segmentation and detection. The sensitivity can be
adjusted using the cutoff.sd parameter, which is an indicator on
how many standard deviations the segment mean value can be before
cutting it off. Will return a warning if start not detected.
}
\examples{
## load a single file
dpa <- load_dpa("data/test.dpa")
## get starting point
start <- dpa_detect_start(dpa, return.plot = TRUE)
## plot the start detection
dpa_detect_start(dpa, return.plot = TRUE)
}
\seealso{
dpa_detect_end, dtrim, dtriml, dtrim_s, dtriml_s
}
