% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{dtrim}
\alias{dtrim}
\title{Automatically trim an individual density profile on both sides}
\usage{
dtrim(dpa, return.plot = FALSE, return.fail = FALSE, silent = FALSE)
}
\arguments{
\item{dpa}{An dpa object, see \code{load_dpa}}

\item{return.plot}{Return a plot instead of dpa object? If TRUE,
returns a plot instead of a dpa object. When return.fail = TRUE,
it returns a list of three: dpa object, start trimming success
and end trimming success.}

\item{return.fail}{Should information on the success of trimming be
included when returning a dpa object?}

\item{silent}{Mute detection warnings, used when calling on list. A
list of trimmed dpa objects, a result of calling dtriml or
dtriml_s on a dpa list with return.fail = FALSE.}
}
\value{
A trimmed dpa object, with the beginning and ending
  removed, if they were detected. When return.plot = TRUE, it
  returns a plot displaying the process.
}
\description{
Calls dpa_detect_start and dpa_detect_end on a given dpa object and
returns a trimmed dpa object with the the row before the starting
point and after the ending removed. If \code{return.plot = TRUE},
it will return a plot displaying the dpa object with detected
starting and ending point. If called with the option
\code{return.fail = FALSE} and \code{return.plot = FALSE}, the
returned object will also include information on whether both
cutoff points were detected. If starting/ending point not detected,
dpa object is returned with no changes. When running on a list of
dpa objects, use \code{dtriml}.
}
\examples{
\dontrun{
## load a single dpa file
dpa  <- load_dpa(system.file("extdata", "00010001.dpa", package = "densiter"))
## trim the measurements
dpa.trimmed <- dtrim(dpa)
## plot trimming
dtrim(dpa, return.plot = TRUE)
}

}
\seealso{
dtriml, dtrim_s, dtriml_s
}
