% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{dtrim_s}
\alias{dtrim_s}
\title{Automatically trim an individual density profile on the starting side}
\usage{
dtrim_s(dpa, return.plot = FALSE, return.fail = FALSE, silent = FALSE)
}
\arguments{
\item{dpa}{An dpa object, see \code{load_dpa}}

\item{return.plot}{Return a plot instead of dpa object? If TRUE,
returns a plot instead of a dpa object. When return.fail = TRUE,
it returns a list of three: dpa object, start trimming success
and end trimming success.}

\item{return.fail}{Should information on the success of trimming be
included when returning a dpa object?}

\item{silent}{Mute detection warnings, used when calling on list. A
list of trimmed dpa objects, a result of calling dtriml or
dtriml_s on a dpa list with return.fail = FALSE.}
}
\value{
A trimmed dpa object, with the beginning and ending
  removed, if they were detected. When return.plot = TRUE, it
  returns a plot displaying the process.
}
\description{
Calls dpa_detect_start on a given dpa object and returns a trimmed
dpa object with the the row before the starting point removed. If
return.plot = TRUE, it will return a plot displaying the dpa object
with detected starting point. If called with the option
return.fail = FALSE and return.plot = FALSE, the returned
object will also include information on whether starting cutoff
point was detected. If starting point not detected, dpa object is
returned with no changes. When running on a list of dpa objects,
use \code{dtriml_s}.
}
\examples{
## load a single file
dpa.list <- load_dpa(file = "test.dpa")
## trim the measurements
dpa.trimmed <- dtrim(dpa.list)
## plot trimming
dtrim(dpa.list, return.plot = TRUE)
}
\seealso{
dtrim, dtriml, dtrim_s
}
