% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{dtrim_sl}
\alias{dtrim_sl}
\title{Automatically trim a list of density profiles on the starting side}
\usage{
dtrim_sl(dpa.list, rreport = FALSE, cl = 1)
}
\arguments{
\item{dpa.list}{A list of dpa objects, see \code{load_dpa}}

\item{rreport}{Return an embedded report on automatic trim success,
mandatory when using \code{correct_failures} to manually pick
starting/ending.}

\item{cl}{Number of cores to run the trimming in parallel, passed through to \code{pbapply}.
starting/ending.}
}
\value{
A list of trimmed dpa objects. When rreport = TRUE, it return a two-item list of (i) trimmed dpa objects and (ii) trimming report data frame.
}
\description{
Calls \code{dtrim} on a list of dpa objects and returns a list of
trimmed objects. If automatic detection fails, the dpa objects are
not trimmed. Can be run in parallel on multiple cores, this speeds
up the trimming process significantly. Only trims the starting
side, see \code{dtriml} for trimming both side simultaneously.
}
\examples{
## load several dpa files
dpa.list <- load_dpa(dpa.directory = system.file("extdata", package = "densiter"))
## trim the measurements
dpa.trimmed <- dtrim_sl(dpa.list)
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
