% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{extract_dpa_name}
\alias{extract_dpa_name}
\title{Extract a file name from a full path}
\usage{
extract_dpa_name(string)
}
\arguments{
\item{string}{A path to file, including file name. Can be nested in
many directories or in none.}
}
\value{
An extracted filename, a character string.
}
\description{
A wrapper function for regex extraction of filename. Given a
character string ("data/0005/00/00050060.dpa"), it will return only
the file name without the extension ("00050060").
}
\examples{
extract_dpa_name("data/0005/00/00050060.dpa")
}
\references{
https://stackoverflow.com/questions/47678725/how-to-do-str-extract-with-base-r
}
