% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dpa}
\alias{load_dpa}
\title{Load a single Resistograph measurement file (*.dpa) or a directory
of *.dpa files.}
\usage{
load_dpa(dpa.file = NULL, dpa.directory = "", recursive = TRUE, name = "file")
}
\arguments{
\item{dpa.file}{A path to a single file, including file name.}

\item{dpa.directory}{A directory with .dpa files.}

\item{recursive}{Also look for dpa files in subfolders?}

\item{name}{Either \code{c("file", "folder")}, used for naming of
list items. If "file", only file name withouth the complete path
will be used for naming ("00050060"). If "folder", the complete
path along with file name will be used to name the dpa objects
("data/0005/00/00050060"). *.dpa ending is removed from the name
in both cases.}
}
\value{
A \code{dpa} object or a list of \code{dpa} objects.
}
\description{
Loads either a single .dpa file or a list of .dpa files. If
dpa.file is specified, it will load a single file. If dpa.directory
is specified, it will search for all dpa files in that directory
(recursively in all subfolders, can be turned off) and return a
list of dpa files. It will use pbapply to display progress, if
loading a directory.
}
\examples{
## load a single file
load_dpa("data/test.dpa")
dpa <- load_dpa("data/0005/00/00050060.dpa")
## load all files in directory
dpa.list <- load_dpa(dpa.directory = "data")
}
\seealso{
read_dpa.
}
