% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all}
\alias{plot_all}
\title{Plot a list of dp objects, one by one}
\usage{
plot_all(dp.list)
}
\arguments{
\item{dp.list}{A list of dp objects, see \code{dpload}}
}
\description{
Plot a list of dp objects, one by one. Press any key to move to
the next dp object. Returns nothing.
}
\examples{
## load several dp files
dp.list <- dpload(dp.directory = system.file("extdata", package = "densiter"))
## trim the measurements
\dontrun{
plot_all(dp.list)
}
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
