% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_failures}
\alias{plot_failures}
\title{Plot trimming failures one by one}
\usage{
plot_failures(dpa.trimmed)
}
\arguments{
\item{dpa.trimmed}{A list of trimmed dpa objects, see \code{load_dpa}}
}
\description{
Plot each failed trimming detection, one by one. Press any key to
move to the next dpa object. Returns nothing. The entry list of dpa
trimmed objects must include the trimming report (rreport = TRUE).
}
\examples{
## load a single file
dpa.list <- load_dpa(folder = "data")
## trim the measurements
plot_failures(dpa.list)
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
