% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_start_detection}
\alias{plot_start_detection}
\title{Display start detection on a list of dpa objects}
\usage{
plot_start_detection(dpa.list)
}
\arguments{
\item{dpa.list}{A list of dpa objects, see \code{load_dpa}}
}
\description{
Display an automatic start detection of dpa list, each dpa object
individually. Press any key to move to the next dpa object. Returns
nothing.
}
\examples{
## load a single file
dpa.list <- load_dpa(folder = "data")
## trim the measurements
plot_start_detection(dpa.list)
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
