% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trimming}
\alias{plot_trimming}
\title{Display automatic trimming on a list of dpa objects}
\usage{
plot_trimming(dpa.list)
}
\arguments{
\item{dpa.list}{A list of dpa objects, see \code{load_dpa}}
}
\description{
Display an automatic trimming of dpa list, each dpa object
individually. Press any key to move to the next dpa object. Returns
nothing.
}
\examples{
## load several dpa files
dpa.list <- load_dpa(dpa.directory = system.file("extdata", package = "densiter"))
## plot trimming the measurements
\dontrun{
plot_trimming(dpa.list)
}
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
