% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_dpa}
\alias{read_dpa}
\title{Read a single Resistograph measurement file (*.dpa)}
\usage{
read_dpa(file)
}
\arguments{
\item{file}{A path to file, including file name.}
}
\value{
A \code{dpa} object.
}
\description{
Reads a single *.dpa file and returns a \code{dpa} object,
constructed from two lists: \code{data} and \code{footer}. The
former one contains actual measurement values, the latter includes
supplementary data recorded by the Resistograph device, such as
time, firmware number...
}
\examples{
read_dpa("data/test.dpa")
}
\seealso{
read_dpa.
}
