% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{splitAt}
\alias{splitAt}
\title{Split a data frame by row numbers, return a list of data frames}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{A data frame.}

\item{pos}{A vector of row number where to split the data frame.A data frame.
embedded ("return.fail = FALSE").}
}
\description{
Found somewhere on StackExchange, posibly on https://
stackoverflow.com/questions/16357962/r-split-numeric-vector-at-
position.
}
