% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{combine_data}
\alias{combine_data}
\title{Combines density measurement from a dpa object list into a single
data frame}
\usage{
combine_data(dpa.list)
}
\arguments{
\item{dpa.list}{A list of dpa objects, either from loading several
files using load_dpa or combined manually. Note: the list should
include only dpa objects!}
}
\value{
A tibble, combining all density data from dpa.list
}
\description{
Given a dpa object list, this function will extract all density
measurement data from all dpa objects in a given list and combine
them in a single data frame/tibble.
}
\examples{
## load all files in directory
dpa.list <- load_dpa(dpa.directory = "data")
combine_data(dpa.list)
}
\seealso{
load_dpa, combine_footer.
}
