% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{correct_failures}
\alias{correct_failures}
\title{Manually correct failures after automatic trim detection}
\usage{
correct_failures(dpa.trimmed)
}
\arguments{
\item{dpa.trimmed}{A list of dpa objects, trimmed, with the report
embedded ("return.fail = FALSE").}
}
\description{
This function will take a list of trimmed dpa objects (a result of
dtriml or dtriml_s function) and interactively ask the user to
assign starting/ending points manually. Follow-up to automatic
trim functions or to be used manually, will display a plot with the
density profiles for each failure in trim detection sequentially.
The plot title will display whether you are selection start or end
positions. Use your mouse to select starting/ending point on the
plot, your selection will then be displayed on the plot. Use keys
y/n to confirm selection, pressing n will restart the selection
process. The selection can be repeated once after pressing n, next
n will stop the process and it has to be restarted manually. Will
return a complete list, both with the non-failed automatically
trimmed dpa objects and those corrected manually. The automatic
trim functions should be called with the option
"return.fail = FALSE", which embeds a trimming report when
returning the list of trimmed dpa objects.
}
\examples{
## load all dpa files
dpa <- load_dpa(folder = "data")
## trim the list
dpa.trimmed <- dtrim(dpa, "return.fail = FALSE")
## manually correct the failures
dpa.corrected <- correct_failures(dpa.trimmed)
}
\seealso{
dtrim, dtriml, manual_trim_detect
}
