% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{manual_trim_detect}
\alias{manual_trim_detect}
\title{Manually select a starting or ending location of a density profile}
\usage{
manual_trim_detect(failure, label = "")
}
\arguments{
\item{failure}{A dp object, usually see dpload.}

\item{label}{Optional label to be displayed on the plot after the
file}
}
\value{
The x position selected on the graph, row number in the dpa$data data frame.
}
\description{
Follow-up to automatic trim functions or to be used manually, will
display a plot with the density profiles. Most commonly used in
automatic failure corrections by the function correct_failures. Use
your mouse to select starting/ending point on the plot, your
selection will then be displayed on the plot. Use keys y/n to
confirm selection, pressing n will restart the selection process.
The selection can be repeated once after pressing n, next n will
stop the process and it has to be restarted manually. This function
uses \code{grDevices::getGraphicsEvent} and
\code{graphics::locator}, which only works on screen devices X11,
windows and quartz. It will not work on other devices.
}
\examples{
## load a single file
dp  <- dpload(system.file("extdata", "00010001.dpa", package = "densiter"))
## get a starting point on the plot
\dontrun{
manual_trim_detect(dp)
}
}
\seealso{
dptrim, dptriml
}
