% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all}
\alias{plot_all}
\title{Plot a list of dpa objects, one by one}
\usage{
plot_all(dpa.list)
}
\arguments{
\item{dpa.list}{A list of dpa objects, see \code{load_dpa}}
}
\description{
Plot a list of dpa objects, one by one. Press any key to move to
the next dpa object. Returns nothing.
}
\examples{
## load several dpa files
dpa.list <- load_dpa(dpa.directory = system.file("extdata", package = "densiter"))
## trim the measurements
\dontrun{
plot_all(dpa.list)
}
}
\seealso{
dtrim, dtrim_s, dtriml_s,
}
