% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_end_detection}
\alias{plot_end_detection}
\title{Display end detection on a list of dp objects}
\usage{
plot_end_detection(dp.list)
}
\arguments{
\item{dp.list}{A list of dp objects, see \code{dpload}}
}
\description{
Display an automatic end detection of dp list, each dp object
individually. Press any key to move to the next dp object. Returns
nothing.
}
\examples{
## load several dp files
dp.list <- dpload(dp.directory = system.file("extdata", package = "densiter"))
## trim the measurements
\dontrun{
plot_end_detection(dp.list)
}
}
\seealso{
dptrim, dptrim_s, dptriml_s,
}
