% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_failures}
\alias{plot_failures}
\title{Plot trimming failures one by one}
\usage{
plot_failures(dp.trimmed)
}
\arguments{
\item{dp.trimmed}{A list of trimmed dp objects, see \code{dpload}}
}
\description{
Plot each failed trimming detection, one by one. Press any key to
move to the next dp object. Returns nothing. The entry list of dp
trimmed objects must include the trimming report (rreport = TRUE).
}
\examples{
## load several dp files
dp.list <- dpload(dp.directory = system.file("extdata", package = "densiter"))
## trim the measurements
\dontrun{
plot_failures(dp.list)
}
}
\seealso{
dptrim, dptrim_s, dptriml_s,
}
