% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trimming}
\alias{plot_trimming}
\title{Display automatic trimming on a list of dp objects}
\usage{
plot_trimming(dp.list)
}
\arguments{
\item{dp.list}{A list of dp objects, see \code{dpload}}
}
\description{
Display an automatic trimming of dp list, each dp object
individually. Press any key to move to the next dp object. Returns
nothing.
}
\examples{
## load several dp files
dp.list <- dpload(dp.directory = system.file("extdata", package = "densitr"))
## plot trimming the measurements
\donttest{
if(interactive()){
plot_trimming(dp.list)
}}
}
\seealso{
dptrim, dptrim_s, dptriml_s,
}
