% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_dpa}
\alias{read_dpa}
\title{Read a single Resistograph measurement file (*.dpa)}
\usage{
read_dpa(file)
}
\arguments{
\item{file}{A path to file, including file name.}
}
\value{
A \code{dp} object.
}
\description{
Reads a single *.dpa file and returns a \code{dp} object,
constructed from two lists: \code{data} and \code{footer}. The
former one contains actual measurement values, the latter includes
supplementary data recorded by the Resistograph device, such as
time, firmware number...
}
\examples{
\dontrun{
read_dpa(system.file("extdata", "00010001.dpa", package = "densiter"))
}
}
\seealso{
dpload
}
