% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{remove_trim_failures}
\alias{remove_trim_failures}
\title{Remove automatic trim failures from a list of trimmed dpa objects
and return only non-failed trimmed objects}
\usage{
remove_trim_failures(dpa.trimmed)
}
\arguments{
\item{dpa.trimmed}{A list of trimmed dpa objects, a result of
calling dtriml or dtriml_s on a dpa list with
"return.fail = FALSE".}
}
\value{
A dpa list of trimmed objects with failures removed.
}
\description{
Returns a dpa list of trimmed dpa objects without the failed trim
objects. Trimmed dpa list should be a result of either calling
dtriml on a list of dpa objects or calling dtriml_s to remove the
starting portions of the measurement. Both functions should be
called with the option "return.fail = FALSE", which embeds a
trimming report when returning the list of trimmed dpa objects.
}
\examples{
## load a single file
dpa.list <- load_dpa(folder = "data")
## trim the measurements
dpa.trimmed <- dtriml(dpa.list, "return.fail = FALSE")
## remove trimming failures
dpa.nofailures <- remove_trim_failures(dpa.trimmed)
}
\seealso{
dtriml, dtriml_s
}
