% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{separate_trim_failures}
\alias{separate_trim_failures}
\title{Remove automatic trim failures from a list of trimmed dpa objects
and return ONLY failures}
\usage{
separate_trim_failures(dpa.trimmed)
}
\arguments{
\item{dpa.trimmed}{A list of trimmed dpa objects, a result of
calling dtriml or dtriml_s on a dpa list with
"return.fail = FALSE".}
}
\value{
Two lists, one with start failures and one with end failures.
}
\description{
An inverse of remove_dpa_failures, return a list of failed trimming
objects from a trimmed dpa list. Trimmed dpa list should be a
result of either calling dtriml on a list of dpa objects or calling
dtriml_s to remove the starting portions of the measurement. Both
functions should be called with the option "rreport = FALSE", which
embeds a trimming report when returning the list of trimmed dpa
objects. If no failures found, it will return a list of trimmed
profiles without the report attached.
}
\examples{
\dontrun{
## load several dpa files
dpa.list <- load_dpa(dpa.directory = system.file("extdata", package = "densiter"))
## trim the measurements
dpa.trimmed <- dtriml(dpa.list, rreport = TRUE)
## separate trimming failures
dpa.nofailures <- separate_trim_failures(dpa.trimmed)
}
}
\seealso{
dtriml, dtriml_s
}
